requirements.txt
PyYAML==6.0.1
python-dotenv==1.0.1
arxiv==2.1.0
beautifulsoup4==4.12.3
requests==2.31.0
PyMuPDF==1.24.7
groq==0.9.0
tqdm==4.66.4
tenacity==8.4.1
langchain-text-splitters==0.2.1 